Fixing butterknife issues
