+++ sage-android/Sage/app/src/main/res/layout/verify_user_filter.xml
@@ -1,13 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <merge xmlns:android="http://schemas.android.com/apk/res/android">
     <LinearLayout
-        android:id="@+id/announcement_filter_container"
+        android:id="@+id/list_filter_container"
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:gravity="center"
         android:background="@color/black72"
-        android:clickable="true">
+        android:clickable="true"
+        android:visibility="gone">
 
         <LinearLayout
             android:layout_width="250dp"
